requirements-dev.txt
anyio==4.9.0
certifi==2025.4.26
charset-normalizer==3.4.2
coverage==7.8.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pytest==8.3.5
pytest-cov==6.1.1
python-dotenv==1.1.0
requests==2.32.3
requests-mock==1.12.1
sniffio==1.3.1
urllib3==2.4.0
